named packages like Ethan's example, reorganized files better
