--- euphoria-service/src/main/java/edu/cooper/ece366/euphoria/Utils.java
@@ -1,17 +1,0 @@
-package edu.cooper.ece366.euphoria;
-
-enum Location {
-    NEWYORK, LONDON, HONGKONG, BERLIN, BEIJING, WASHINGTON
-}
-
-enum EducationLevel {
-    NOHIGHSCHOOL, HIGHSCHOOL, GED, SOMECOLLEGE, ASSOCIATES, BACHELORS, MASTERS, PHD, MD, JD
-}
-
-enum SkillLevel {
-    INTERNSHIP, ENTRYLEVEL, ASSOCIATE, SENIOR, DIRECTOR, EXECUTIVE
-}
-
-enum Industry {
-    EDUCATION, ENERGY, FINANCE, FOOD, HEALTHCARE, INSURANCE, MEDIA, RETAIL, SERVICES, TECHNOLOGY, TRANSPORT, UTILITIES
-}
