+++ euphoria-service/src/main/java/edu/cooper/ece366/euphoria/handler/PostingHandlers.java
@@ -1,4 +1,9 @@
-package edu.cooper.ece366.euphoria;
+package edu.cooper.ece366.euphoria.handler;
+
+import edu.cooper.ece366.euphoria.model.*;
+import edu.cooper.ece366.euphoria.utils.Industry;
+import edu.cooper.ece366.euphoria.utils.Location;
+import edu.cooper.ece366.euphoria.utils.SkillLevel;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.annotations.VisibleForTesting;
