+++ euphoria-service/src/main/java/edu/cooper/ece366/euphoria/handler/UserHandlers.java
@@ -1,4 +1,7 @@
-package edu.cooper.ece366.euphoria;
+package edu.cooper.ece366.euphoria.handler;
+
+import edu.cooper.ece366.euphoria.model.*;
+import edu.cooper.ece366.euphoria.utils.EducationLevel;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.annotations.VisibleForTesting;
