+++ euphoria-service/src/main/java/edu/cooper/ece366/euphoria/store/UserStore.java
@@ -1,0 +1,9 @@
+package edu.cooper.ece366.euphoria.Store;
+
+import edu.cooper.ece366.euphoria.model.User;
+
+public interface UserStore {
+
+    User getBook(String id);
+
+}
