+++ euphoria-service/src/main/java/edu/cooper/ece366/euphoria/store/UserStoreJdbc.java
@@ -1,0 +1,4 @@
+package edu.cooper.ece366.euphoria.Store;
+
+public class UserStoreJdbc {
+}
