+++ euphoria-service/src/main/java/edu/cooper/ece366/euphoria/utils/EducationLevel.java
@@ -1,0 +1,6 @@
+package edu.cooper.ece366.euphoria.utils;
+
+public enum EducationLevel {
+    NOHIGHSCHOOL, HIGHSCHOOL, GED, SOMECOLLEGE, ASSOCIATES, BACHELORS, MASTERS, PHD, MD, JD
+}
+
