+++ euphoria-service/src/main/java/edu/cooper/ece366/euphoria/utils/Location.java
@@ -1,0 +1,5 @@
+package edu.cooper.ece366.euphoria.utils;
+
+public enum Location {
+    NEWYORK, LONDON, HONGKONG, BERLIN, BEIJING, WASHINGTON
+}
