+++ euphoria-service/src/main/java/edu/cooper/ece366/euphoria/utils/SkillLevel.java
@@ -1,0 +1,5 @@
+package edu.cooper.ece366.euphoria.utils;
+
+public enum SkillLevel {
+    INTERNSHIP, ENTRYLEVEL, ASSOCIATE, SENIOR, DIRECTOR, EXECUTIVE
+}
